chore: migrating java-sdk-common to this repo (#15)
This PR copies existing java-sdk-common code from the private repo and
takes the same approach as the java-server-sdk-otel package for having
it use github actions and release-please for CICD.

I made a separate commit for the code copy to make that easier to
verify. All code in lib/shared/common should be identical to the
[existing repo
](https://github.com/launchdarkly/java-sdk-common-private) with no
changes intended.